[Loosely] Enforces IComparable. And cleaned Condition Attribute
Weird bug  that if I create a condition and the toggle is set to true for a bool return type, the results is false.

In addition, strictly enforcing IComparable is bugging out. Suppose returnType is System.Int32, it says it is not an IComparable even though it is and it works in the code, it does not work in the EditorWindow